requirements_dev.txt
Django==2.2.15  #core
pytz==2020.1    #core
sqlparse==0.3.1 #core
pillow==7.2.0   #core - images
django-ckeditor==5.9.0  #blog
django-js-asset==1.2.2  #ckeditor - blog
django-taggit==1.3.0    #blog
